+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## 6.3.2 - 2026-09-23
 
 ### @igojs/server